:recycle: refactor: login package name change auth
